Extracted MouseClickConverter to separate class
